Separated the code signing dossier tool into a "reader" and a "writer".
PiperOrigin-RevId: 486996527
(cherry picked from commit 6fa02bfa2fb6812fa42421aef4c7c3e275a2ab21)
